disable test-rwlock1
It is a bad test because it cannot link to pthread, but the package builds
properly and the compiled tools are working properly.

It is added to Makefile.in, not Makefile.am because it won't get reconfigured.
A new patch or a dh_autoreconf will break it, but that's fine, because this
test also might get fixed and we want to know about that.